moved grpchelpers to common, added bearer token support to grpc requests
